Dodano dwie aktywności
Dodano wstępnie aktywność wyświetlania szczegółów celów
Dodano wstępnie aktywność edycji nazwy i opisu celu

Zmieniono sposób odświeżania ViewPager2 z fragmentami celów

Przygotowano struktury plików dla Intencji Wdrożeniowych

Dodano wstępnie obsługę menu

Dodano tłumaczenia

Poprawki, refaktoring

+++ app/src/main/java/com/example/towardsgoalsapp/goals/GoalData.kt
@@ -1,7 +1,7 @@
 package com.example.towardsgoalsapp.goals
 
 data class GoalData(
-    val goalId: Int,
+    val goalId: Long,
     val goalName: String,
     val goalDescription: String,
     val progress: Double
